First Pass at Basic Functionality for 2021x
Filling out structure of variation points between pre 2021x and 2021x plus in Player Piano. This does not include actually completing the sections that vary between those versions; the only sections known to be implemented for 2021x are those that handle the supplied patterns (and did not cause model issues when ran previously). For those supplied patterns, this version of Player Piano replicates what the pre2021x version of Player Piano did